pyproject.toml
[project]
name = "aoe-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index>=0.12.34",
    "python-dotenv>=1.1.0",
]
